Move Join and Disconnect events into PlayerEvent
