+++ server/patches/server/network/ServerGamePacketListenerImpl.patch
@@ -36,7 +36,7 @@
 -        this.server.getPlayerList().broadcastMessage((new TranslatableComponent("multiplayer.player.left", new Object[]{this.player.getDisplayName()})).withStyle(ChatFormatting.YELLOW), ChatType.SYSTEM, Util.NIL_UUID);
 +        // Loom start :: fire ConnectionEvent.Disconnect
 +        var message = (new TranslatableComponent("multiplayer.player.left", new Object[]{this.player.getDisplayName()})).withStyle(ChatFormatting.YELLOW);
-+        var event = org.loomdev.loom.event.LoomEventDispatcher.fire(new org.loomdev.loom.event.server.connection.ConnectionEventImpl.DisconnectImpl(player, message));
++        var event = org.loomdev.loom.event.LoomEventDispatcher.fire(new org.loomdev.loom.event.entity.PlayerEventImpl.DisconnectImpl(player, message));
 +        if (event.getDisconnectMessage().toString().length() != 0) {
 +            this.server.getPlayerList().broadcastMessage(org.loomdev.loom.util.transformer.TextTransformer.toMinecraft(event.getDisconnectMessage()), ChatType.SYSTEM, Util.NIL_UUID);
 +        }
