+++ server/src/main/java/org/loomdev/loom/event/entity/PlayerEventImpl.java
@@ -4,15 +4,13 @@ import net.kyori.adventure.text.Component;
 import net.minecraft.server.level.ServerPlayer;
 import org.jetbrains.annotations.NotNull;
 import org.loomdev.api.entity.player.Player;
-import org.loomdev.api.event.Cancelable;
 import org.loomdev.api.event.entity.PlayerEvent;
 import org.loomdev.loom.event.EventImpl;
 import org.loomdev.loom.util.transformer.TextTransformer;
 
-public class PlayerEventImpl extends EventImpl implements PlayerEvent, Cancelable {
+public class PlayerEventImpl extends EventImpl implements PlayerEvent {
 
     private final Player player;
-    private boolean canceled;
 
     public PlayerEventImpl(Player player) {
         this.player = player;
@@ -24,19 +22,10 @@ public class PlayerEventImpl extends EventImpl implements PlayerEvent, Cancelabl
         return player;
     }
 
-    @Override
-    public boolean isCanceled() {
-        return canceled;
-    }
-
-    @Override
-    public void setCanceled(boolean canceled) {
-        this.canceled = canceled;
-    }
-
     public static class ChatImpl extends PlayerEventImpl implements PlayerEvent.Chat {
 
         private Component message;
+        private boolean canceled;
 
         public ChatImpl(ServerPlayer player, net.minecraft.network.chat.Component message) {
             super((Player) player.getLoomEntity());
@@ -51,6 +40,58 @@ public class PlayerEventImpl extends EventImpl implements PlayerEvent, Cancelabl
 
         @Override
         public void setMessage(@NotNull Component component) {
+            this.message = component;
+        }
+
+        @Override
+        public boolean isCanceled() {
+            return canceled;
+        }
+
+        @Override
+        public void setCanceled(boolean canceled) {
+            this.canceled = canceled;
+        }
+    }
+
+    public static class JoinImpl extends PlayerEventImpl implements PlayerEvent.Join {
+
+        private Component message;
+
+        public JoinImpl(ServerPlayer player, net.minecraft.network.chat.Component component) {
+            super((Player) player.getLoomEntity());
+            this.message = TextTransformer.toLoom(component);
+        }
+
+        @Override
+        @NotNull
+        public Component getJoinMessage() {
+            return message;
+        }
+
+        @Override
+        public void setJoinMessage(@NotNull Component component) {
+            this.message = component;
+        }
+    }
+
+    public static class DisconnectImpl extends PlayerEventImpl implements PlayerEvent.Disconnect {
+
+        private Component message;
+
+        public DisconnectImpl(ServerPlayer player, net.minecraft.network.chat.Component component) {
+            super((Player) player.getLoomEntity());
+            this.message = TextTransformer.toLoom(component);
+        }
+
+        @Override
+        @NotNull
+        public Component getDisconnectMessage() {
+            return message;
+        }
+
+        @Override
+        public void setDisconnectMessage(@NotNull Component component) {
             this.message = component;
         }
     }
